# docs/swagger.json
{"swagger":"2.0","info":{"contact":{}},"paths":{"/auth/jwt/update":{"post":{"consumes":["application/json"],"produces":["application/json"],"tags":["jwt"],"summary":"Update Access JWT token","parameters":[{"type":"string","description":"User's refresh token","name":"Authorization","in":"header","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/controllers.SingleAccessToken"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/errors.PublicPCCError"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/errors.PublicPCCError"}}}}},"/cart/":{"get":{"consumes":["application/json"],"produces":["application/json"],"tags":["cart"],"summary":"Get user's cart","parameters":[{"type":"string","description":"access token for authorization","name":"Authorization","in":"header","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/models.Cart"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/errors.PublicPCCError"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/errors.PublicPCCError"}},"403":{"description":"Forbidden","schema":{"$ref":"#/definitions/errors.PublicPCCError"}}}},"put":{"consumes":["application/json"],"produces":["application/json"],"tags":["cart"],"summary":"Update product's quantity","parameters":[{"type":"string","description":"access token for authorization","name":"Authorization","in":"header","required":true},{"description":"info about a product to add","name":"obj","in":"body","required":true,"schema":{"$ref":"#/definitions/inputs.AddToCartInput"}}],"responses":{"200":{"description":"OK","schema":{"type":"integer"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/errors.PublicPCCError"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/errors.PublicPCCError"}},"403":{"description":"Forbidden","schema":{"$ref":"#/definitions/errors.PublicPCCError"}}}},"post":{"consumes":["application/json"],"produces":["application/json"],"tags":["cart"],"summary":"Add product to a cart","parameters":[{"type":"string","description":"access token for authorization","name":"Authorization","in":"header","required":true},{"description":"info about a product to add","name":"obj","in":"body","required":true,"schema":{"$ref":"#/definitions/inputs.AddToCartInput"}}],"responses":{"201":{"description":"Created","schema":{"type":"integer"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/errors.PublicPCCError"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/errors.PublicPCCError"}},"403":{"description":"Forbidden","schema":{"$ref":"#/definitions/errors.PublicPCCError"}}}},"delete":{"consumes":["application/json"],"produces":["application/json"],"tags":["cart"],"summary":"Remove the product from the cart","parameters":[{"type":"string","description":"access token for authorization","name":"Authorization","in":"header","required":true},{"description":"info about a product to remove","name":"obj","in":"body","required":true,"schema":{"$ref":"#/definitions/inputs.RemoveFromCartInput"}}],"responses":{"200":{"description":"OK","schema":{"type":"integer"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/errors.PublicPCCError"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/errors.PublicPCCError"}},"403":{"description":"Forbidden","schema":{"$ref":"#/definitions/errors.PublicPCCError"}}}}},"/categories/":{"get":{"consumes":["application/json"],"produces":["application/json"],"tags":["categories"],"summary":"Get all categories","responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/models.Category"}}},"400":{"description":"Bad Request","schema":{"type":"object","additionalProperties":true}}}}},"/laptops/add":{"post":{"consumes":["application/json"],"produces":["application/json"],"tags":["laptops"],"summary":"Add a new laptop","parameters":[{"description":"Laptop data","name":"laptop","in":"body","required":true,"schema":{"$ref":"#/definitions/inputs.AddLaptopInput"}},{"type":"string","description":"access token for user with Admin role","name":"Authorization","in":"header","required":true}],"responses":{"201":{"description":"Created","schema":{"type":"object","additionalProperties":true}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/errors.PublicPCCError"}},"403":{"description":"Forbidden","schema":{"$ref":"#/definitions/errors.PublicPCCError"}}}}},"/laptops/chars/{id}":{"get":{"consumes":["application/json"],"produces":["application/json"],"tags":["laptops"],"summary":"Get laptop chars","parameters":[{"type":"integer","description":"Laptop id","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/models.LaptopChars"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/errors.PublicPCCError"}}}}},"/products/":{"get":{"consumes":["application/json"],"produces":["application/json"],"tags":["products"],"summary":"Get products from page N in quantity M","parameters":[{"type":"integer","name":"count","in":"query"},{"type":"integer","name":"page","in":"query"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/models.Product"}}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/errors.PublicPCCError"}}}}},"/products/{id}":{"get":{"consumes":["application/json"],"produces":["application/json"],"tags":["products"],"summary":"Get a single product by ID","parameters":[{"type":"integer","description":"Product ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/models.Product"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/errors.PublicPCCError"}}}}},"/users/login":{"get":{"consumes":["application/json"],"produces":["application/json"],"tags":["users"],"summary":"Login","parameters":[{"description":"User data to login","name":"user","in":"body","required":true,"schema":{"$ref":"#/definitions/inputs.LoginUserInput"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/outputs.JWTPair"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/errors.PublicPCCError"}}}}},"/users/register":{"post":{"consumes":["application/json"],"produces":["application/json"],"tags":["users"],"summary":"Register a new User","parameters":[{"description":"User data to register","name":"user","in":"body","required":true,"schema":{"$ref":"#/definitions/inputs.RegisterUserInput"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/models.User"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/errors.PublicPCCError"}}}}}},"definitions":{"controllers.SingleAccessToken":{"type":"object","properties":{"access_token":{"type":"string"}}},"errors.ErrorCode":{"type":"integer","enum":[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22],"x-enum-varnames":["EC_INTERNAL","EC_DB_UNIQUE_FAIL","EC_DB_NOT_FOUND_ERROR","EC_DB_WRONG_CHARS_TABLE_NAME","EC_DB_ROLE_ERROR","EC_DB_OTHER","EC_CTRLS_INPUT_ERROR","EC_CTRLS_NO_USER_DATA_ERROR","EC_REDIS_NIL","EC_REDIS_OTHER","EC_REDIS_WRONG_VALUE","EC_JSON_MARSHAL_ERROR","EC_JSON_UNMARSHAL_ERROR","EC_JWT_TOKEN_MALFORMED","EC_JWT_TOKEN_EXPIRED","EC_JWT_TOKEN_NOT_VALID_YET","EC_UNKNOWN_JWT_TOKEN_ERROR","EC_JWT_ERROR_TOKEN_TYPE","EC_HEADER_MISSING","EC_ATOI_NOT_AN_UINT","EC_ATOI_RANGE_ERROR","EC_ATOI_UNKNOWN","EC_ROLE_LOWER"]},"errors.ErrorKind":{"type":"string","enum":["internal","database","jwt","controllers","redis","json","header","atoi","roles"],"x-enum-varnames":["EK_INTERNAL","EK_DATABASE","EK_JWT","EK_CTRLS","EK_REDIS","EK_JSON","EK_HEADER","EK_ATOI","EK_ROLES"]},"errors.PublicPCCError":{"type":"object","properties":{"code":{"description":"Code represents the error code in terms of this project\nIt should NEVER contain the inner code from the database drive or something similar","allOf":[{"$ref":"#/definitions/errors.ErrorCode"}]},"details":{"description":"Details may provide more details about the error but should NEVER contain the secret or\nsensetive data"},"kind":{"description":"Kind represents the information about the module or the step on which the error occured","allOf":[{"$ref":"#/definitions/errors.ErrorKind"}]},"message":{"description":"SafeMessage contains the summary message that should be safe","type":"string"}}},"inputs.AddLaptopInput":{"type":"object","properties":{"cpu":{"type":"string"},"gpu":{"type":"string"},"name":{"type":"string"},"price":{"type":"number"},"ram":{"type":"integer"},"stock":{"type":"integer"}}},"inputs.AddToCartInput":{"type":"object","properties":{"product_id":{"type":"integer"},"quantity":{"type":"integer"}}},"inputs.LoginUserInput":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string"},"password":{"type":"string"}}},"inputs.RegisterUserInput":{"type":"object","required":["email","name","password"],"properties":{"email":{"type":"string"},"name":{"type":"string"},"password":{"type":"string"}}},"inputs.RemoveFromCartInput":{"type":"object","properties":{"product_id":{"type":"integer"}}},"models.Cart":{"type":"object","properties":{"id":{"type":"integer"},"items":{"type":"array","items":{"$ref":"#/definitions/models.CartItem"}},"user_id":{"type":"integer"}}},"models.CartItem":{"type":"object","properties":{"added_at":{"type":"string"},"id":{"type":"integer"},"product":{"$ref":"#/definitions/models.Product"},"quantity":{"type":"integer"}}},"models.Category":{"type":"object","properties":{"description":{"type":"string"},"icon":{"type":"string"},"id":{"type":"integer"},"slug":{"type":"string"},"title":{"type":"string"}}},"models.LaptopChars":{"type":"object","properties":{"cpu":{"type":"string"},"gpu":{"type":"string"},"id":{"type":"integer"},"ram":{"type":"integer"}}},"models.Product":{"type":"object","properties":{"id":{"type":"integer"},"name":{"type":"string"},"price":{"type":"number"},"selled":{"type":"integer"},"stock":{"type":"integer"}}},"models.User":{"type":"object","properties":{"email":{"type":"string"},"id":{"type":"integer"},"name":{"type":"string"},"user_role":{"$ref":"#/definitions/models.UserRole"}}},"models.UserRole":{"type":"string","enum":["Temporary","Default","Admin"],"x-enum-varnames":["Temporary","Default","Admin"]},"outputs.JWTPair":{"type":"object","properties":{"access_token":{"type":"string"},"refresh_token":{"type":"string"}}}}}